Refactor: Restructure project into a package-based monorepo.
This commit completes a major refactoring to transition the project to a more robust and scalable monorepo structure.

Key Changes:

1.  **Package-based Structure:**
    - The `bot` service in `packages/bot/` has been converted into a self-contained, installable Python package.
    - All of the bot's source code has been moved into a `packages/bot/src/` directory to enforce a clean separation between source and configuration.
    - A dedicated `packages/bot/pyproject.toml` now manages the bot's specific dependencies and entry points.
    - The root `pyproject.toml` has been simplified to manage only shared development tools.

2.  **Hybrid Configuration:**
    - The configuration loader (`loader.py`) has been updated to support a hybrid strategy.
    - It now loads a shared `.env` file from the project root for common variables and a package-specific `.env` for secrets like `DISCORD_TOKEN`.
    - Path logic was updated to function correctly within the new `src` layout.

3.  **Generalized Dockerfile:**
    - The root `Dockerfile` has been modified to be reusable for any package in the monorepo.
    - It now accepts a `PACKAGE_NAME` build argument to specify which package to build, making the CI/CD process more flexible.

4.  **Testing and Cleanup:**
    - All import paths in the application and test suite have been updated to match the new structure.
    - All 49 tests pass, verifying the correctness of the refactoring.
    - Build artifacts (`.egg-info`) have been removed and added to `.gitignore` to prevent them from being committed.
