+++ packages/bot/pyproject.toml
@@ -1,0 +1,44 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+packages = {find = {where = ["src"]}}
+
+[project]
+name = "ryuuko-bot"
+version = "1.3.1"
+authors = [
+  { name="Zang Vũ", email="zvwgvx@polydevs.uk" },
+]
+description = "A Discord bot."
+readme = "README.md"
+requires-python = ">=3.11"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "python-dotenv>=0.21.0",
+    "discord.py>=2.3.2",
+    "requests>=2.28.0",
+    "pymongo>=4.0.0",
+    "Pillow>=10.0.0",
+    "aiohttp>=3.9.0",
+    "urllib3>=1.21.0",
+    "tiktoken>=0.4.0",
+    "colorama>=0.4.6",
+    "google-genai>=0.4.0"
+]
+
+[project.optional-dependencies]
+test = [
+    "pytest>=7.0.0",
+    "pytest-asyncio>=0.21.0",
+    "pytest-mock>=3.10.0"
+]
+
+[project.scripts]
+start = "bot.__main__:main"
