--- packages/bot/ryuuko_chatbot.egg-info/SOURCES.txt
@@ -1,41 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-src/__init__.py
-src/__main__.py
-src/bot/__init__.py
-src/bot/main.py
-src/bot/api/__init__.py
-src/bot/api/client.py
-src/bot/commands/__init__.py
-src/bot/commands/admin.py
-src/bot/commands/basic.py
-src/bot/commands/system.py
-src/bot/commands/user.py
-src/bot/events/__init__.py
-src/bot/events/messages.py
-src/bot/services/__init__.py
-src/bot/services/auth.py
-src/config/__init__.py
-src/config/loader.py
-src/config/user_config.py
-src/llm_services/main.py
-src/llm_services/providers/__init__.py
-src/llm_services/providers/_common.py
-src/llm_services/providers/aistudio.py
-src/llm_services/providers/polydevs.py
-src/llm_services/providers/proxyvn.py
-src/ryuuko_chatbot.egg-info/PKG-INFO
-src/ryuuko_chatbot.egg-info/SOURCES.txt
-src/ryuuko_chatbot.egg-info/dependency_links.txt
-src/ryuuko_chatbot.egg-info/entry_points.txt
-src/ryuuko_chatbot.egg-info/requires.txt
-src/ryuuko_chatbot.egg-info/top_level.txt
-src/storage/__init__.py
-src/storage/database.py
-src/storage/memory.py
-src/utils/__init__.py
-src/utils/health.py
-src/utils/logger.py
-src/utils/queue.py
-src/utils/ratelimit.py
