--- packages/bot/ryuuko_chatbot.egg-info/entry_points.txt
@@ -1,2 +1,0 @@
-[console_scripts]
-start = src.__main__:main
