--- packages/bot/ryuuko_chatbot.egg-info/requires.txt
@@ -1,17 +1,0 @@
-python-dotenv>=0.21.0
-discord.py>=2.3.2
-requests>=2.28.0
-pymongo>=4.0.0
-Pillow>=10.0.0
-aiohttp>=3.9.0
-urllib3>=1.21.0
-tiktoken>=0.4.0
-colorama>=0.4.6
-google-genai>=0.4.0
-fastapi>=0.118.0
-uvicorn>=0.37.0
-
-[test]
-pytest>=7.0.0
-pytest-asyncio>=0.21.0
-pytest-mock>=3.10.0
