--- packages/bot/ryuuko_chatbot.egg-info/top_level.txt
@@ -1,7 +1,0 @@
-__init__
-__main__
-bot
-config
-llm_services
-storage
-utils
