+++ packages/bot/src/bot/__init__.py
@@ -7,8 +7,7 @@ __author__ = "Zang Vũ"
 
 # Main packages
 from . import config
-from . import bot
 from . import storage
 from . import utils
 
-__all__ = ['config', 'bot', 'storage', 'utils']
+__all__ = ['config', 'storage', 'utils']
