+++ packages/bot/src/bot/__main__.py
