+++ packages/bot/src/bot/commands/__init__.py
