+++ packages/bot/src/bot/commands/system.py
