+++ packages/bot/src/bot/commands/user.py
