+++ packages/bot/src/bot/events/__init__.py
