+++ packages/bot/src/bot/llm_services/providers/_common.py
