+++ packages/bot/src/bot/llm_services/providers/aistudio.py
