+++ packages/bot/src/bot/llm_services/providers/instructions.json
