+++ packages/bot/src/bot/main.py
