+++ packages/bot/src/bot/services/auth.py
