+++ packages/bot/src/bot/storage/memory.py
