--- packages/ryuuko_bot.egg-info/PKG-INFO
@@ -1,115 +1,0 @@
-Metadata-Version: 2.4
-Name: ryuuko-bot
-Version: 1.3.1
-Summary: A Discord bot.
-Author-email: Zang Vũ <zvwgvx@polydevs.uk>
-Classifier: Programming Language :: Python :: 3
-Classifier: License :: OSI Approved :: MIT License
-Classifier: Operating System :: OS Independent
-Requires-Python: >=3.11
-Description-Content-Type: text/markdown
-License-File: LICENSE
-Requires-Dist: python-dotenv>=0.21.0
-Requires-Dist: discord.py>=2.3.2
-Requires-Dist: requests>=2.28.0
-Requires-Dist: pymongo>=4.0.0
-Requires-Dist: Pillow>=10.0.0
-Requires-Dist: aiohttp>=3.9.0
-Requires-Dist: urllib3>=1.21.0
-Requires-Dist: tiktoken>=0.4.0
-Requires-Dist: colorama>=0.4.6
-Requires-Dist: google-genai>=0.4.0
-Requires-Dist: fastapi>=0.118.0
-Requires-Dist: uvicorn>=0.37.0
-Provides-Extra: test
-Requires-Dist: pytest>=7.0.0; extra == "test"
-Requires-Dist: pytest-asyncio>=0.21.0; extra == "test"
-Requires-Dist: pytest-mock>=3.10.0; extra == "test"
-Dynamic: license-file
-
-# Ryuuko — Discord LLM Bot
-
-**Ryuuko** is a modular and extensible Discord bot powered by Large Language Models (LLMs). It's designed to be a versatile chatbot that can be customized and extended with new commands and functionalities. The project is structured as a monorepo to support future expansion with other services like a web interface or an API.
-
-## Features
-
-*   **Modular Architecture**: Easily extend the bot by adding new commands, events, or services.
-*   **LLM Integration**: Connects with LLM providers for intelligent conversation.
-*   **Conversation Memory**: Remembers previous messages in a conversation for better context.
-*   **Configuration Management**: Flexible configuration system using environment variables and a `config.json` file.
-*   **Role-Based Access Control**: Restrict commands to authorized users.
-*   **Asynchronous Processing**: Uses a request queue to handle LLM API calls without blocking the bot.
-
-## Getting Started
-
-These instructions will get you a copy of the project up and running on your local machine for development and testing purposes.
-
-### Prerequisites
-
-*   Python 3.11+
-*   MongoDB server
-*   A Discord Bot Token
-
-### Installation
-
-1.  **Clone the repository:**
-    ```bash
-    git clone https://github.com/zvwgvx/ryuuko-chatbot.git
-    cd ryuuko-chatbot
-    ```
-
-2.  **Install dependencies:**
-    The project is structured as a monorepo. The bot is the first package.
-    ```bash
-    # Install main dependencies for the bot
-    pip install .
-
-    # To install testing dependencies as well
-    pip install .[test]
-    ```
-
-3.  **Configure the bot:**
-    *   Create a `.env` file in the root directory and add your secrets. For a full list of required API keys, see the [Setup Guide](docs/SETUP.md).
-        ```env
-        # Discord and Database
-        DISCORD_TOKEN="your_discord_bot_token"
-        MONGODB_CONNECTION_STRING="your_mongodb_connection_string"
-
-        # LLM API Keys (only add the ones you use)
-        AISTUDIO_API_KEY="your_aistudio_api_key"
-        POLYDEVS_API_KEY="your_polydevs_api_key"
-        PROXYVN_API_KEY="your_proxyvn_api_key"
-        ```
-
-    *   Modify `config.json` for additional settings if needed.
-
-## Usage
-
-To run the bot, use one of the following commands:
-
-```bash
-# Using the python module directly
-python -m bot
-
-# Using the installed script
-start
-```
-
-Once the bot is running, you can interact with it on Discord using the `.` prefix. For a full list of commands, see the [Commands Reference](docs/COMMANDS.md).
-
-## Documentation
-
-For more detailed information, please refer to the following documents:
-
-*   [**Architecture**](docs/ARCHITECTURE.md): An overview of the project's technical design.
-*   [**Setup Guide**](docs/SETUP.md): Detailed setup and configuration instructions.
-*   [**Command Reference**](docs/COMMANDS.md): A complete list of all available bot commands.
-*   [**Deployment Guide**](docs/DEPLOYMENT.md): Instructions for deploying the bot to a production environment.
-
-## Contributing
-
-We welcome contributions! Please see our [Contributing Guidelines](CONTRIBUTING.md) to get started.
-
-## License
-
-This project is licensed under the MIT License - see the [LICENSE](LICENSE) file for details.
