+++ pyproject.toml
@@ -2,37 +2,19 @@
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools]
-packages = ["bot"]
-package-dir = {"" = "packages"}
-
 [project]
-name = "ryuuko-bot"
+name = "ryuuko-workspace"
 version = "1.3.1"
-authors = [
-  { name="Zang Vũ", email="zvwgvx@polydevs.uk" },
-]
-description = "A Discord bot."
-readme = "README.md"
+description = "A monorepo for the Ryuuko project."
 requires-python = ">=3.11"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-]
+# This file is for managing shared development tools and dependencies.
+# Packages inside packages/ have their own pyproject.toml files.
+
+# Dependencies for packages that are not yet self-contained.
+# TODO: Move these to their respective packages (e.g., packages/api)
 dependencies = [
-    "python-dotenv>=0.21.0",
-    "discord.py>=2.3.2",
-    "requests>=2.28.0",
-    "pymongo>=4.0.0",
-    "Pillow>=10.0.0",
-    "aiohttp>=3.9.0",
-    "urllib3>=1.21.0",
-    "tiktoken>=0.4.0",
-    "colorama>=0.4.6",
-    "google-genai>=0.4.0",
     "fastapi>=0.118.0",
-    "uvicorn>=0.37.0",
+    "uvicorn>=0.37.0"
 ]
 
 [project.optional-dependencies]
@@ -41,6 +23,3 @@ test = [
     "pytest-asyncio>=0.21.0",
     "pytest-mock>=3.10.0"
 ]
-
-[project.scripts]
-start = "bot.__main__:main"
